utils/requirements.txt
typer
python-dotenv
ruff
psycopg2-binary
requests
SQLAlchemy